SM_backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
httpx
python-multipart
